protein_localization/requirements.txt
# Core dependencies for Protein Sub-Cellular Localization Pipeline
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image processing
tifffile>=2021.7.2
scikit-image>=0.18.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
torch-geometric>=2.0.0

# Graph processing
networkx>=2.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Jupyter support
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
scikit-learn>=0.24.0
